requirements.txt
twilio==8.10.0
flask==3.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
geopy==2.4.1
stripe==7.10.0
pyjwt==2.8.0
bcrypt==4.1.1
phonenumbers==8.13.25
